items_by_index works, proper pks stored
