+++ packages/storage-plus/src/legacy_helpers.rs
@@ -25,23 +25,6 @@ pub(crate) fn length_prefixed_with_key(namespace: &[u8], key: &[u8]) -> Vec<u8> 
     out
 }
 
-/// This is equivalent concat(to_length_prefixed_nested(namespaces), key)
-/// But more efficient when the intermediate namespaces often must be recalculated
-pub(crate) fn namespaces_with_key(namespaces: &[&[u8]], key: &[u8]) -> Vec<u8> {
-    let mut size = key.len();
-    for &namespace in namespaces {
-        size += namespace.len() + 2;
-    }
-
-    let mut out = Vec::with_capacity(size);
-    for &namespace in namespaces {
-        out.extend_from_slice(&encode_length(namespace));
-        out.extend_from_slice(namespace);
-    }
-    out.extend_from_slice(key);
-    out
-}
-
 // pub(crate) fn decode_length(prefix: [u8; 2]) -> usize {
 pub(crate) fn decode_length(prefix: &[u8]) -> usize {
     // TODO: enforce exactly 2 bytes somehow, but usable with slices
