+++ packages/storage-plus/src/map.rs
@@ -33,10 +33,6 @@ where
     T: Serialize + DeserializeOwned,
     K: PrimaryKey<'a>,
 {
-    pub fn root(&'a self) -> &'a [u8] {
-        self.namespace
-    }
-
     pub fn key(&self, k: K) -> Path<T> {
         Path::new(self.namespace, &k.key())
     }
